Got rid of the temp files. I made the changes into weekly and still uses temporary mock data and functions but they work. Added some new buttons to add CRUD actions like adding, deleting, and modifying the habits. Then I also added a button to save any changes into the habits with x and checkmarks so that once it is connected to a db it doesnt do too many requests and stuff. The UI looks good and now we just need to be able to add, delete, and modify the habits but that should be on something different like different issues for the github repo. Also made a change for the date format added a function to change the date format.
