+++ apps/web/src/App.tsx
@@ -2,8 +2,7 @@ import { Routes, Route } from "react-router-dom";
 import MainLayout from "./layouts/MainLayout";
 import Home from "./pages/home";
 import Weekly from "./pages/weekly"
-import WeekViewPage from "./pages/temp";
-import WeekTViewPage from "./pages/weeklyT";
+
 
 export default function App() {
   return (
@@ -11,8 +10,6 @@ export default function App() {
       <Route path="/" element={<MainLayout />}>
         <Route index element={<Home />} />
         <Route path="weekly" element={<Weekly />} />
-        <Route path="temp" element={<WeekViewPage/>} />
-        <Route path="weeklyT" element={<WeekTViewPage/>} />
       </Route>
     </Routes>
   );
